Move StatementProtocolConfig from network service to light-base
